fix: BGE-294 address chat PR review — fix data race and ring buffer eviction
Blocking fixes:
- Merge ChatRepository reads into ChatRepositoryWrite behind the same lock,
  eliminating the data race between concurrent reads and ring-buffer writes.
  Delete standalone ChatRepository; ChatController now injects only
  ChatRepositoryWrite for both reads and writes.
- GetMessagesAfter: when the supplied ID is not found (evicted from ring buffer),
  fall back to a full reload (TakeLast(50)) instead of returning empty, so
  slow clients do not silently stop receiving messages.

Nit fixes:
- Remove dead @using BrowserGameEngine.BlazorClient.Code import from Chat.razor
- Remove unused GameId route parameter and @page "/games/{GameId}/chat" route
- Remove dead chatWindowRef ElementReference field

Tests updated: GetMessagesAfter_UnknownId_ReturnsEmpty renamed to
GetMessagesAfter_EvictedId_FallsBackToFullReload and updated to assert
the full reload behaviour.
